split host_world_manager into manager & writer parts
