+++ shared/src/world/host/host_world_manager.rs
@@ -7,14 +7,11 @@ use std::{
 };
 
 use crate::{
-    messages::channels::senders::indexed_message_writer::IndexedMessageWriter,
     sequence_list::SequenceList,
     world::{
         entity::entity_converters::GlobalWorldManagerType, local_world_manager::LocalWorldManager,
     },
-    BitWrite, BitWriter, ComponentKind, ComponentKinds, ConstBitLength, DiffMask, EntityAction,
-    EntityActionType, EntityConverterMut, Instant, LocalEntityConverter, MessageIndex, PacketIndex,
-    Serde, UnsignedVariableInteger, WorldRefType,
+    ComponentKind, DiffMask, EntityAction, Instant, MessageIndex, PacketIndex,
 };
 
 use super::{entity_action_event::EntityActionEvent, world_channel::WorldChannel};
@@ -28,18 +25,27 @@ pub type ActionId = MessageIndex;
 /// sync on the Client
 pub struct HostWorldManager<E: Copy + Eq + Hash + Send + Sync> {
     // World
-    world_channel: WorldChannel<E>,
+    pub world_channel: WorldChannel<E>,
 
     // Actions
-    next_send_actions: VecDeque<(ActionId, EntityActionEvent<E>)>,
-    sent_action_packets: SequenceList<(Instant, Vec<(ActionId, EntityAction<E>)>)>,
+    pub sent_action_packets: SequenceList<(Instant, Vec<(ActionId, EntityAction<E>)>)>,
 
     // Updates
-    next_send_updates: HashMap<E, HashSet<ComponentKind>>,
     /// Map of component updates and [`DiffMask`] that were written into each packet
-    sent_updates: HashMap<PacketIndex, (Instant, HashMap<(E, ComponentKind), DiffMask>)>,
+    pub sent_updates: HashMap<PacketIndex, (Instant, HashMap<(E, ComponentKind), DiffMask>)>,
     /// Last [`PacketIndex`] where a component update was written by the server
-    last_update_packet_index: PacketIndex,
+    pub last_update_packet_index: PacketIndex,
+}
+
+pub struct HostWorldEvents<E: Copy + Eq + Hash + Send + Sync> {
+    pub next_send_actions: VecDeque<(ActionId, EntityActionEvent<E>)>,
+    pub next_send_updates: HashMap<E, HashSet<ComponentKind>>,
+}
+
+impl<E: Copy + Eq + Hash + Send + Sync> HostWorldEvents<E> {
+    pub fn has_events(&self) -> bool {
+        !self.next_send_actions.is_empty() || !self.next_send_updates.is_empty()
+    }
 }
 
 impl<E: Copy + Eq + Hash + Send + Sync> HostWorldManager<E> {
@@ -51,11 +57,9 @@ impl<E: Copy + Eq + Hash + Send + Sync> HostWorldManager<E> {
         HostWorldManager {
             // World
             world_channel: WorldChannel::new(address, global_world_manager),
-            next_send_actions: VecDeque::new(),
             sent_action_packets: SequenceList::new(),
 
             // Update
-            next_send_updates: HashMap::new(),
             sent_updates: HashMap::new(),
             last_update_packet_index: 0,
         }
@@ -106,17 +110,9 @@ impl<E: Copy + Eq + Hash + Send + Sync> HostWorldManager<E> {
 
     // Messages
 
-    pub fn collect_outgoing_messages(&mut self, now: &Instant, rtt_millis: &f32) {
+    pub fn collect_outgoing_messages(&mut self, rtt_millis: &f32) {
         self.collect_dropped_update_packets(rtt_millis);
-
         self.collect_dropped_action_packets();
-        self.collect_next_actions(now, rtt_millis);
-
-        self.collect_component_updates();
-    }
-
-    pub fn has_outgoing_messages(&self) -> bool {
-        !self.next_send_actions.is_empty() || !self.next_send_updates.is_empty()
     }
 
     // Collecting
@@ -138,10 +134,6 @@ impl<E: Copy + Eq + Hash + Send + Sync> HostWorldManager<E> {
                 return;
             }
         }
-    }
-
-    fn collect_next_actions(&mut self, now: &Instant, rtt_millis: &f32) {
-        self.next_send_actions = self.world_channel.take_next_actions(now, rtt_millis);
     }
 
     fn collect_dropped_update_packets(&mut self, rtt_millis: &f32) {
@@ -197,477 +189,11 @@ impl<E: Copy + Eq + Hash + Send + Sync> HostWorldManager<E> {
         }
     }
 
-    fn collect_component_updates(&mut self) {
-        self.next_send_updates = self.world_channel.collect_next_updates();
-    }
-
-    // Writing actions
-
-    fn write_action_id(
-        writer: &mut dyn BitWrite,
-        last_id_opt: &mut Option<ActionId>,
-        current_id: &ActionId,
-    ) {
-        IndexedMessageWriter::write_message_index(writer, last_id_opt, current_id);
-        *last_id_opt = Some(*current_id);
-    }
-
-    pub fn write_actions<W: WorldRefType<E>>(
-        &mut self,
-        component_kinds: &ComponentKinds,
-        now: &Instant,
-        writer: &mut BitWriter,
-        packet_index: &PacketIndex,
-        world: &W,
-        global_world_manager: &dyn GlobalWorldManagerType<E>,
-        local_world_manager: &mut LocalWorldManager<E>,
-        has_written: &mut bool,
-    ) {
-        let mut last_counted_id: Option<MessageIndex> = None;
-        let mut last_written_id: Option<MessageIndex> = None;
-
-        loop {
-            if self.next_send_actions.is_empty() {
-                break;
-            }
-
-            // check that we can write the next message
-            let mut counter = writer.counter();
-            self.write_action(
-                component_kinds,
-                world,
-                global_world_manager,
-                local_world_manager,
-                packet_index,
-                &mut counter,
-                &mut last_counted_id,
-                false,
-            );
-
-            if counter.overflowed() {
-                // if nothing useful has been written in this packet yet,
-                // send warning about size of component being too big
-                if !*has_written {
-                    self.warn_overflow_action(
-                        component_kinds,
-                        global_world_manager,
-                        counter.bits_needed(),
-                        writer.bits_free(),
-                    );
-                }
-                break;
-            }
-
-            *has_written = true;
-
-            // write ActionContinue bit
-            true.ser(writer);
-
-            // optimization
-            if !self
-                .sent_action_packets
-                .contains_scan_from_back(packet_index)
-            {
-                self.sent_action_packets
-                    .insert_scan_from_back(*packet_index, (now.clone(), Vec::new()));
-            }
-
-            // write data
-            self.write_action(
-                component_kinds,
-                world,
-                global_world_manager,
-                local_world_manager,
-                packet_index,
-                writer,
-                &mut last_written_id,
-                true,
-            );
-
-            // pop action we've written
-            self.next_send_actions.pop_front();
+    pub fn take_outgoing_events(&mut self, now: &Instant, rtt_millis: &f32) -> HostWorldEvents<E> {
+        HostWorldEvents {
+            next_send_actions: self.world_channel.take_next_actions(now, rtt_millis),
+            next_send_updates: self.world_channel.collect_next_updates(),
         }
-    }
-
-    #[allow(clippy::too_many_arguments)]
-    fn write_action<W: WorldRefType<E>>(
-        &mut self,
-        component_kinds: &ComponentKinds,
-        world: &W,
-        global_world_manager: &dyn GlobalWorldManagerType<E>,
-        local_world_manager: &mut LocalWorldManager<E>,
-        packet_index: &PacketIndex,
-        writer: &mut dyn BitWrite,
-        last_written_id: &mut Option<ActionId>,
-        is_writing: bool,
-    ) {
-        let (action_id, action) = self.next_send_actions.front().unwrap();
-
-        // write message id
-        Self::write_action_id(writer, last_written_id, action_id);
-
-        match action {
-            EntityActionEvent::SpawnEntity(world_entity) => {
-                EntityActionType::SpawnEntity.ser(writer);
-
-                // write net entity
-                local_world_manager
-                    .entity_to_local_entity(world_entity)
-                    .unwrap()
-                    .host_ser(writer);
-
-                // get component list
-                let component_kind_list = match global_world_manager.component_kinds(world_entity) {
-                    Some(kind_list) => kind_list,
-                    None => Vec::new(),
-                };
-
-                // write number of components
-                let components_num =
-                    UnsignedVariableInteger::<3>::new(component_kind_list.len() as i128);
-                components_num.ser(writer);
-
-                for component_kind in &component_kind_list {
-                    let mut converter = EntityConverterMut::new(
-                        global_world_manager.to_global_entity_converter(),
-                        local_world_manager,
-                    );
-
-                    // write component payload
-                    world
-                        .component_of_kind(world_entity, component_kind)
-                        .expect("Component does not exist in World")
-                        .write(component_kinds, writer, &mut converter);
-                }
-
-                // if we are writing to this packet, add it to record
-                if is_writing {
-                    Self::record_action_written(
-                        &mut self.sent_action_packets,
-                        packet_index,
-                        action_id,
-                        EntityAction::SpawnEntity(*world_entity, component_kind_list),
-                    );
-                }
-            }
-            EntityActionEvent::DespawnEntity(world_entity) => {
-                EntityActionType::DespawnEntity.ser(writer);
-
-                // write net entity
-                local_world_manager
-                    .entity_to_local_entity(world_entity)
-                    .unwrap()
-                    .host_ser(writer);
-
-                // if we are writing to this packet, add it to record
-                if is_writing {
-                    Self::record_action_written(
-                        &mut self.sent_action_packets,
-                        packet_index,
-                        action_id,
-                        EntityAction::DespawnEntity(*world_entity),
-                    );
-                }
-            }
-            EntityActionEvent::InsertComponent(world_entity, component) => {
-                if !world.has_component_of_kind(world_entity, component)
-                    || !self.world_channel.entity_channel_is_open(world_entity)
-                {
-                    EntityActionType::Noop.ser(writer);
-
-                    // if we are actually writing this packet
-                    if is_writing {
-                        // add it to action record
-                        Self::record_action_written(
-                            &mut self.sent_action_packets,
-                            packet_index,
-                            action_id,
-                            EntityAction::Noop,
-                        );
-                    }
-                } else {
-                    EntityActionType::InsertComponent.ser(writer);
-
-                    // write net entity
-                    local_world_manager
-                        .entity_to_local_entity(world_entity)
-                        .unwrap()
-                        .host_ser(writer);
-
-                    let mut converter = EntityConverterMut::new(
-                        global_world_manager.to_global_entity_converter(),
-                        local_world_manager,
-                    );
-
-                    // write component payload
-                    world
-                        .component_of_kind(world_entity, component)
-                        .expect("Component does not exist in World")
-                        .write(component_kinds, writer, &mut converter);
-
-                    // if we are actually writing this packet
-                    if is_writing {
-                        // add it to action record
-                        Self::record_action_written(
-                            &mut self.sent_action_packets,
-                            packet_index,
-                            action_id,
-                            EntityAction::InsertComponent(*world_entity, *component),
-                        );
-                    }
-                }
-            }
-            EntityActionEvent::RemoveComponent(world_entity, component_kind) => {
-                if !self.world_channel.entity_channel_is_open(world_entity) {
-                    EntityActionType::Noop.ser(writer);
-
-                    // if we are actually writing this packet
-                    if is_writing {
-                        // add it to action record
-                        Self::record_action_written(
-                            &mut self.sent_action_packets,
-                            packet_index,
-                            action_id,
-                            EntityAction::Noop,
-                        );
-                    }
-                } else {
-                    EntityActionType::RemoveComponent.ser(writer);
-
-                    // write net entity
-                    local_world_manager
-                        .entity_to_local_entity(world_entity)
-                        .unwrap()
-                        .host_ser(writer);
-
-                    // write component kind
-                    component_kind.ser(component_kinds, writer);
-
-                    // if we are writing to this packet, add it to record
-                    if is_writing {
-                        Self::record_action_written(
-                            &mut self.sent_action_packets,
-                            packet_index,
-                            action_id,
-                            EntityAction::RemoveComponent(*world_entity, *component_kind),
-                        );
-                    }
-                }
-            }
-        }
-    }
-
-    #[allow(clippy::type_complexity)]
-    fn record_action_written(
-        sent_actions: &mut SequenceList<(Instant, Vec<(ActionId, EntityAction<E>)>)>,
-        packet_index: &PacketIndex,
-        action_id: &ActionId,
-        action_record: EntityAction<E>,
-    ) {
-        let (_, sent_actions_list) = sent_actions.get_mut_scan_from_back(packet_index).unwrap();
-        sent_actions_list.push((*action_id, action_record));
-    }
-
-    fn warn_overflow_action(
-        &self,
-        component_kinds: &ComponentKinds,
-        global_world_manager: &dyn GlobalWorldManagerType<E>,
-        bits_needed: u32,
-        bits_free: u32,
-    ) {
-        let (_action_id, action) = self.next_send_actions.front().unwrap();
-
-        match action {
-            EntityActionEvent::SpawnEntity(entity) => {
-                let component_kind_list = match global_world_manager.component_kinds(entity) {
-                    Some(kind_list) => kind_list,
-                    None => Vec::new(),
-                };
-
-                let mut component_names = "".to_owned();
-                let mut added = false;
-
-                for component_kind in &component_kind_list {
-                    if added {
-                        component_names.push(',');
-                    } else {
-                        added = true;
-                    }
-                    let name = component_kinds.kind_to_name(component_kind);
-                    component_names.push_str(&name);
-                }
-                panic!(
-                    "Packet Write Error: Blocking overflow detected! Entity Spawn message with Components `{component_names}` requires {bits_needed} bits, but packet only has {bits_free} bits available! Recommend slimming down these Components."
-                )
-            }
-            EntityActionEvent::InsertComponent(_entity, component_kind) => {
-                let component_name = component_kinds.kind_to_name(component_kind);
-                panic!(
-                    "Packet Write Error: Blocking overflow detected! Component Insertion message of type `{component_name}` requires {bits_needed} bits, but packet only has {bits_free} bits available! This condition should never be reached, as large Messages should be Fragmented in the Reliable channel"
-                )
-            }
-            _ => {
-                panic!(
-                    "Packet Write Error: Blocking overflow detected! Action requires {bits_needed} bits, but packet only has {bits_free} bits available! This message should never display..."
-                )
-            }
-        }
-    }
-
-    pub fn write_updates<W: WorldRefType<E>>(
-        &mut self,
-        component_kinds: &ComponentKinds,
-        now: &Instant,
-        writer: &mut BitWriter,
-        packet_index: &PacketIndex,
-        world: &W,
-        global_world_manager: &dyn GlobalWorldManagerType<E>,
-        local_world_manager: &mut LocalWorldManager<E>,
-        has_written: &mut bool,
-    ) {
-        let all_update_entities: Vec<E> = self.next_send_updates.keys().copied().collect();
-
-        for entity in all_update_entities {
-            // check that we can at least write a LocalEntity and a ComponentContinue bit
-            let mut counter = writer.counter();
-
-            // get LocalEntity
-            let local_entity = local_world_manager.entity_to_local_entity(&entity).unwrap();
-
-            // write LocalEntity
-            local_entity.host_ser(&mut counter);
-            counter.write_bit(false);
-
-            if counter.overflowed() {
-                break;
-            }
-
-            // write UpdateContinue bit
-            true.ser(writer);
-
-            // reserve ComponentContinue bit
-            writer.reserve_bits(1);
-
-            // write LocalEntity
-            local_entity.host_ser(writer);
-
-            // write Components
-            self.write_update(
-                component_kinds,
-                now,
-                world,
-                global_world_manager,
-                local_world_manager,
-                packet_index,
-                writer,
-                &entity,
-                has_written,
-            );
-
-            // write ComponentContinue finish bit, release
-            false.ser(writer);
-            writer.release_bits(1);
-        }
-    }
-
-    /// For a given entity, write component value updates into a packet
-    /// Only component values that changed in the internal (naia's) host world will be written
-    fn write_update<W: WorldRefType<E>>(
-        &mut self,
-        component_kinds: &ComponentKinds,
-        now: &Instant,
-        world: &W,
-        global_world_manager: &dyn GlobalWorldManagerType<E>,
-        local_world_manager: &mut LocalWorldManager<E>,
-        packet_index: &PacketIndex,
-        writer: &mut BitWriter,
-        entity: &E,
-        has_written: &mut bool,
-    ) {
-        let mut written_component_kinds = Vec::new();
-        let component_kind_set = self.next_send_updates.get(entity).unwrap();
-        for component_kind in component_kind_set {
-            // get diff mask
-            let diff_mask = self
-                .world_channel
-                .diff_handler
-                .diff_mask(entity, component_kind)
-                .expect("DiffHandler does not have registered Component!")
-                .clone();
-
-            let mut converter = EntityConverterMut::new(
-                global_world_manager.to_global_entity_converter(),
-                local_world_manager,
-            );
-
-            // check that we can write the next component update
-            let mut counter = writer.counter();
-            counter.write_bits(<ComponentKind as ConstBitLength>::const_bit_length());
-            world
-                .component_of_kind(entity, component_kind)
-                .expect("Component does not exist in World")
-                .write_update(&diff_mask, &mut counter, &mut converter);
-
-            if counter.overflowed() {
-                // if nothing useful has been written in this packet yet,
-                // send warning about size of component being too big
-                if !*has_written {
-                    let component_name = component_kinds.kind_to_name(component_kind);
-                    self.warn_overflow_update(
-                        component_name,
-                        counter.bits_needed(),
-                        writer.bits_free(),
-                    );
-                }
-
-                break;
-            }
-
-            *has_written = true;
-
-            // write ComponentContinue bit
-            true.ser(writer);
-
-            // write component kind
-            component_kind.ser(component_kinds, writer);
-
-            // write data
-            world
-                .component_of_kind(entity, component_kind)
-                .expect("Component does not exist in World")
-                .write_update(&diff_mask, writer, &mut converter);
-
-            written_component_kinds.push(*component_kind);
-
-            // place diff mask in a special transmission record - like map
-            self.last_update_packet_index = *packet_index;
-
-            if !self.sent_updates.contains_key(packet_index) {
-                self.sent_updates
-                    .insert(*packet_index, (now.clone(), HashMap::new()));
-            }
-            let (_, sent_updates_map) = self.sent_updates.get_mut(packet_index).unwrap();
-            sent_updates_map.insert((*entity, *component_kind), diff_mask);
-
-            // having copied the diff mask for this update, clear the component
-            self.world_channel
-                .diff_handler
-                .clear_diff_mask(entity, component_kind);
-        }
-
-        let update_kinds = self.next_send_updates.get_mut(entity).unwrap();
-        for component_kind in &written_component_kinds {
-            update_kinds.remove(component_kind);
-        }
-        if update_kinds.is_empty() {
-            self.next_send_updates.remove(entity);
-        }
-    }
-
-    fn warn_overflow_update(&self, component_name: String, bits_needed: u32, bits_free: u32) {
-        panic!(
-            "Packet Write Error: Blocking overflow detected! Data update of Component `{component_name}` requires {bits_needed} bits, but packet only has {bits_free} bits available! Recommended to slim down this Component"
-        )
     }
 }
