+++ shared/src/world/host/mod.rs
@@ -1,5 +1,6 @@
 pub mod global_diff_handler;
 pub mod host_world_manager;
+pub mod host_world_writer;
 pub mod mut_channel;
 pub mod user_diff_handler;
 pub mod world_channel;
